pyproject.toml
[project]
name = "page-index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "faiss-cpu>=1.13.1",
    "litellm>=1.80.0",
    "mistral-common>=1.8.6",
    "mistralai>=1.9.11",
    "openai==1.101.0",
    "pymupdf==1.26.4",
    "pypdf2==3.0.1",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "tiktoken==0.11.0",
]
